requirements.txt
AWSIoTPythonSDK==1.5.4
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
dotenv==0.9.9
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
pymongo==4.12.0
python-dotenv==1.1.0
Werkzeug==3.1.3